feat: Started implementing state machines
Used the state machine pattern to implement some of the existing
behavior in player_character. I didn't get finished though. So, I made a
duplicate of player_character that has my changes. I left the state
machine disabled, but that doesn't seem to 100% stop it. Anyway once it
is finished, it won't matter.