refactor: Created a EffectAudioPlayer to manage volume
The EffectAudioPlayer has one and only one job. To set the effect volume
correctly. It listens for any change to the Global volume and then
updates itself. This should allow us to change the effect volume in the
game, not just in the main menu.

I want to do the same thing with the background player. It would
primarily be responsible for managing the "quite" mode when pause is
selected. It isn't as useful, but would still be.