feat: Break appart Main Menu and Intro
The main menu and intro were coupled together and made it possible to do
things like select the menu behind the intro animation. To fix, I just
broke apart the scenes in two. Because of that, we no longer free the
scene when the animation ends. Instead, we just change scenes on the
tree.

This also makes all the menu buttons a consistent size.

Finally, I think there was supposed to be a transition into the main
menu, but it wasn't working. I got it working, but we can also just
remove it.