refactor: Create a TimedDeathComponent
Trying to make it so that we don't have to keep doing tons of
boilerplate code. In that effort, I found a ton of code that does "death
after timer". So I pulled out a component to do it. I did find one spot
in the code with a timer, but it didn't appear to be used. It was really
weird.